RAG Chatbot for Insurance Agencies/Backend/requirements.txt
fastapi
uvicorn
pydantic
pypdf
tiktoken
faiss-cpu
numpy
openai
reportlab